add ahocorasick
