+++ bukkit/loader/src/main/resources/craft-engine.properties
@@ -26,3 +26,4 @@ adventure-text-minimessage=${adventure_bundle_version}
 adventure-text-serializer-gson=${adventure_bundle_version}
 adventure-text-serializer-json=${adventure_bundle_version}
 netty-codec-http=${netty_version}
+ahocorasick=${ahocorasick_version}
