+++ core/src/main/java/net/momirealms/craftengine/core/item/ItemBuildContext.java
@@ -2,7 +2,7 @@ package net.momirealms.craftengine.core.item;
 
 import net.kyori.adventure.text.minimessage.tag.resolver.TagResolver;
 import net.momirealms.craftengine.core.entity.player.Player;
-import net.momirealms.craftengine.core.plugin.minimessage.*;
+import net.momirealms.craftengine.core.plugin.text.minimessage.*;
 import net.momirealms.craftengine.core.util.context.ContextHolder;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
