+++ core/src/main/java/net/momirealms/craftengine/core/plugin/text/minimessage/I18NTag.java
@@ -1,35 +1,34 @@
-package net.momirealms.craftengine.core.plugin.minimessage;
+package net.momirealms.craftengine.core.plugin.text.minimessage;
 
 import net.kyori.adventure.text.minimessage.Context;
 import net.kyori.adventure.text.minimessage.ParsingException;
 import net.kyori.adventure.text.minimessage.tag.Tag;
 import net.kyori.adventure.text.minimessage.tag.resolver.ArgumentQueue;
 import net.kyori.adventure.text.minimessage.tag.resolver.TagResolver;
-import net.momirealms.craftengine.core.entity.player.Player;
-import net.momirealms.craftengine.core.plugin.CraftEngine;
+import net.momirealms.craftengine.core.plugin.locale.TranslationManager;
 import net.momirealms.craftengine.core.util.AdventureHelper;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public class PlaceholderTag implements TagResolver {
-    private final Player player;
+public class I18NTag implements TagResolver {
+    private final MiniMessageTextContext context;
 
-    public PlaceholderTag(@Nullable Player player) {
-        this.player = player;
+    public I18NTag(MiniMessageTextContext context) {
+        this.context = context;
     }
 
     @Override
     public @Nullable Tag resolve(@NotNull String name, @NotNull ArgumentQueue arguments, @NotNull Context ctx) throws ParsingException {
-        if (!this.has(name) || !CraftEngine.instance().hasPlaceholderAPI()) {
+        if (!this.has(name)) {
             return null;
         }
-        String placeholder = arguments.popOr("No argument placeholder provided").toString();
-        String parsed = CraftEngine.instance().parse(player, "%" + placeholder + "%");
-        return Tag.inserting(AdventureHelper.miniMessage(parsed));
+        String i18nKey = arguments.popOr("No argument i18n key provided").toString();
+        String translation = TranslationManager.instance().miniMessageTranslation(i18nKey);
+        return Tag.inserting(AdventureHelper.miniMessage().deserialize(translation, this.context.tagResolvers()));
     }
 
     @Override
     public boolean has(@NotNull String name) {
-        return "papi".equals(name);
+        return "i18n".equals(name) || "l10n".equals(name);
     }
 }
