+++ core/src/main/java/net/momirealms/craftengine/core/plugin/text/minimessage/PlaceholderTag.java
@@ -1,38 +1,35 @@
-package net.momirealms.craftengine.core.plugin.minimessage;
+package net.momirealms.craftengine.core.plugin.text.minimessage;
 
 import net.kyori.adventure.text.minimessage.Context;
 import net.kyori.adventure.text.minimessage.ParsingException;
 import net.kyori.adventure.text.minimessage.tag.Tag;
 import net.kyori.adventure.text.minimessage.tag.resolver.ArgumentQueue;
 import net.kyori.adventure.text.minimessage.tag.resolver.TagResolver;
+import net.momirealms.craftengine.core.entity.player.Player;
 import net.momirealms.craftengine.core.plugin.CraftEngine;
 import net.momirealms.craftengine.core.util.AdventureHelper;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public class ShiftTag implements TagResolver {
-    public static final ShiftTag INSTANCE = new ShiftTag();
+public class PlaceholderTag implements TagResolver {
+    private final Player player;
 
-    public static ShiftTag instance() {
-        return INSTANCE;
+    public PlaceholderTag(@Nullable Player player) {
+        this.player = player;
     }
 
     @Override
     public @Nullable Tag resolve(@NotNull String name, @NotNull ArgumentQueue arguments, @NotNull Context ctx) throws ParsingException {
-        if (!this.has(name)) {
+        if (!this.has(name) || !CraftEngine.instance().hasPlaceholderAPI()) {
             return null;
         }
-        String shiftAmount = arguments.popOr("No argument shift provided").toString();
-        try {
-            int shift = Integer.parseInt(shiftAmount);
-            return Tag.inserting(AdventureHelper.miniMessage(CraftEngine.instance().imageManager().createMiniMessageOffsets(shift)));
-        } catch (NumberFormatException e) {
-            throw ctx.newException("Invalid shift value", arguments);
-        }
+        String placeholder = arguments.popOr("No argument placeholder provided").toString();
+        String parsed = CraftEngine.instance().parse(player, "%" + placeholder + "%");
+        return Tag.inserting(AdventureHelper.miniMessage(parsed));
     }
 
     @Override
     public boolean has(@NotNull String name) {
-        return "shift".equals(name);
+        return "papi".equals(name);
     }
 }
