+++ gradle.properties
@@ -47,6 +47,7 @@ joml_version=1.10.8
 datafixerupper_version=1.0.20
 mojang_brigadier_version=1.0.18
 byte_buddy_version=1.15.11
+ahocorasick_version=0.6.3
 snake_yaml_version=2.3
 anti_grief_version=0.13
 nms_helper_version=0.34
